backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.23
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
openai==1.12.0
python-multipart==0.0.9
joblib==1.3.2
alembic==1.12.1
pandas==2.1.3
scikit-learn==1.3.2
reportlab==4.0.7
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2
pytest-asyncio==0.21.1
pytest-mock==3.12.0
langchain-core==0.1.40
langchain==0.1.9
langchain-openai==0.0.5
langchain-community==0.0.24
langchain-text-splitters==0.0.1
langchain-chroma==0.1.0
chromadb==0.4.22
tiktoken==0.5.2
pypdf==4.0.2
psycopg2-binary==2.9.9
PyPDF2
requests
aiohttp 